qcacld-3.0: Update firmware filter config (part 2)
In order to conserve firmware memory when various filtering features are
not in use, update the target configuration provided to firmware based
on the intersection of various filtering features and host
configuration.

Part 1
	num_wow_filters = ARP/NS offload enabled ? 2 : 5
	bpf_instruction_size - BPF enabled ? 0 (auto) : 0xffffffff

Part 2
	num_packet_filter = Packet Filter enabled ? 12 : 0xffffffff

CRs-Fixed: 2199297

+++ components/pmo/core/inc/wlan_pmo_pkt_filter.h
@@ -32,7 +32,7 @@ struct wlan_objmgr_psoc;
  *
  * Return: number of packet filters
  */
-uint8_t pmo_get_num_packet_filters(struct wlan_objmgr_psoc *psoc);
+uint32_t pmo_get_num_packet_filters(struct wlan_objmgr_psoc *psoc);
 
 /**
  * pmo_set_pkt_fltr_req() - Set packet filter
